backend/requirement.txt
Django>=4.2
djangorestframework
mysqlclient
python-dotenv
django-cors-headers
Pillow
